refactor: AccountConnection now handles account connection, finally. :art:
